feat: command results can now optionally ignored when creating or cloasing tabs
`auto-tab-groups-create-commands` and `auto-tab-groups-close-commands` may now
include the optional property `:ignore-result`. If it is not nil it alters the
behavior of the advice functions.

When give for a "create" command: If the property is `nil` the tab group is
first created and then the original function is called. Otherwise, the results
are passed to the tab group name function to obtain the name dynamically and
create the group afterwards.

When give for a "close" command: If the property is `nil` the tab group is only
closed if the command returns not nil, otherwise ts result is ignored and the
group is closed anyways.